+++ colour.h
@@ -16,8 +16,8 @@ constexpr int_fast8_t White = 0;
 constexpr int_fast8_t Black = 1;
 constexpr int_fast8_t ColourNb = 2;
 
-constexpr int_fast32_t WhiteFlag = 1 << White;
-constexpr int_fast32_t BlackFlag = 1 << Black;
+constexpr int_fast8_t WhiteFlag = 1 << White;
+constexpr int_fast8_t BlackFlag = 1 << Black;
 
 // macros
